remove old token selector and continue work on refactor
