+++ packages/ui/src/img/AllChainsLogo.tsx
@@ -66,8 +66,8 @@ export const AllChainsLogo = (props: React.HTMLAttributes<SVGElement>) => {
           fill="white"
         />
         <path
-          fill-rule="evenodd"
-          clip-rule="evenodd"
+          fillRule="evenodd"
+          clipRule="evenodd"
           d="M70.2864 257.805C102.922 257.805 129.378 231.349 129.378 198.714C129.378 166.079 102.922 139.623 70.2864 139.623C37.6514 139.623 11.1953 166.079 11.1953 198.714C11.1953 231.349 37.6514 257.805 70.2864 257.805ZM48.0469 214.945C49.6636 215.454 51.4151 215.708 53.3013 215.708C57.6426 215.708 61.1006 214.706 63.6753 212.7C66.25 210.664 68.0464 207.595 69.0644 203.493C69.3639 202.206 69.6483 200.918 69.9177 199.631C70.2172 198.343 70.4715 197.041 70.6812 195.724C71.0404 193.688 70.9806 191.907 70.5015 190.38C70.0524 188.853 69.259 187.565 68.1213 186.517C67.0135 185.47 65.6513 184.691 64.0346 184.182C62.4479 183.643 60.7113 183.374 58.8251 183.374C54.454 183.374 50.9809 184.422 48.4062 186.517C45.8315 188.613 44.05 191.682 43.062 195.724C42.7625 197.041 42.4633 198.343 42.1638 199.631C41.8943 200.918 41.6398 202.206 41.4003 203.493C41.0711 205.529 41.1309 207.31 41.58 208.837C42.0591 210.364 42.8523 211.637 43.9602 212.655C45.068 213.673 46.4302 214.436 48.0469 214.945ZM58.1515 207.804C56.924 208.762 55.5018 209.241 53.8851 209.241C52.2384 209.241 51.0558 208.762 50.3373 207.804C49.6187 206.846 49.4391 205.319 49.7984 203.224C50.0378 201.906 50.2775 200.664 50.5169 199.496C50.7864 198.329 51.0858 197.116 51.4151 195.859C51.924 193.763 52.7773 192.236 53.9749 191.278C55.2024 190.32 56.6246 189.841 58.2413 189.841C59.858 189.841 61.0406 190.32 61.7891 191.278C62.5377 192.236 62.7173 193.763 62.3281 195.859C62.1184 197.116 61.879 198.329 61.6095 199.496C61.3701 200.664 61.0855 201.906 60.7562 203.224C60.2473 205.319 59.3789 206.846 58.1515 207.804ZM71.8001 214.945C71.9798 215.155 72.2192 215.259 72.5187 215.259H78.6263C78.9556 215.259 79.2399 215.155 79.4796 214.945C79.7491 214.735 79.9136 214.466 79.9736 214.137L82.0394 204.301H88.1022C91.9644 204.301 95.0331 203.478 97.3086 201.831C99.6138 200.185 101.141 197.64 101.889 194.197C102.249 192.46 102.233 190.949 101.844 189.661C101.455 188.344 100.767 187.251 99.7786 186.383C98.7906 185.515 97.548 184.871 96.0511 184.452C94.584 184.033 92.9373 183.823 91.1111 183.823H79.1652C78.8658 183.823 78.5814 183.928 78.312 184.137C78.0425 184.347 77.8778 184.616 77.818 184.946L71.6205 214.137C71.5604 214.466 71.6205 214.735 71.8001 214.945ZM88.3716 198.104H83.2071L84.9585 190.065H90.3476C91.4255 190.065 92.2187 190.245 92.7278 190.604C93.2668 190.963 93.5811 191.443 93.6709 192.041C93.7607 192.64 93.7307 193.329 93.5811 194.107C93.2816 195.454 92.638 196.457 91.65 197.116C90.6918 197.775 89.5991 198.104 88.3716 198.104Z"
           fill="#FF0000"
         />
@@ -121,12 +121,12 @@ export const AllChainsLogo = (props: React.HTMLAttributes<SVGElement>) => {
           y2="163.808"
           gradientUnits="userSpaceOnUse"
         >
-          <stop offset="0.08" stop-color="#9945FF" />
-          <stop offset="0.3" stop-color="#8752F3" />
-          <stop offset="0.5" stop-color="#5497D5" />
-          <stop offset="0.6" stop-color="#43B4CA" />
-          <stop offset="0.72" stop-color="#28E0B9" />
-          <stop offset="0.97" stop-color="#19FB9B" />
+          <stop offset="0.08" stopColor="#9945FF" />
+          <stop offset="0.3" stopColor="#8752F3" />
+          <stop offset="0.5" stopColor="#5497D5" />
+          <stop offset="0.6" stopColor="#43B4CA" />
+          <stop offset="0.72" stopColor="#28E0B9" />
+          <stop offset="0.97" stopColor="#19FB9B" />
         </linearGradient>
       </defs>
     </svg>
